+++ viewclust/insta_plot.py
@@ -42,7 +42,7 @@ def insta_plot(clust_info, cores_queued, cores_running, resample_str='',
     jobUse: Generates the input frames for this function.
     """
 
-    print('This function is deprecated as of v0.3.0.\nSupport continues in the ViewClust-Vis package.', file=sys.err)
+    print('This function is deprecated as of v0.3.0.\nSupport continues in the ViewClust-Vis package.', file=sys.stderr)
 
     # Temp vars so that we aren't touching anything by ref
     clust_info_tmp = clust_info.copy()